Implement "listprojects" and "listcontexts" command
Useful to get an overview of all projects / contexts in use.
Supports the same filters as the "list" command.

Note: If a task has multiple contexts like "@person @phone"
and you filter for @phon*, you will also get @person
in the output as the filter selects on the task level only.