Resolve '~' to home directory for 'todo-file' and 'done-file' (#90)
This adds the ability to resolve a '~' used in the 'todo-file' and
'done-file' variable of the configuration to be resolved by the 'dirs'
crate to the home directory. The path is only changed when 'dirs' can
resolve the home directory, so it should be save on windows and
unsupported systems.